requirements.txt
web3==6.15.1
python-dotenv==1.0.1
requests==2.31.0
eth-account==0.11.0

